requirements.txt
playwright==1.41.0
pytest-playwright==0.4.0
python-dotenv==1.0.0
tqdm==4.66.1 